requirements.txt
asgiref==3.3.1
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
Django==3.1.4
django-cors-headers==3.6.0
django-request-logging==0.7.2
httpie==2.3.0
idna==2.10
mysqlclient==2.0.2
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
PySocks==1.7.1
pytz==2020.4
requests==2.25.1
requests-toolbelt==0.9.1
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
wincertstore==0.2
